chore: Update .gitignore to ignore VS Code files

+++ models/mdRoomType.py
@@ -4,7 +4,7 @@ from app import db
 class RoomType(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(50))
-    price_per_night = db.Column(db.Float)
+    price_per_night = db.Column(db.DECIMAL)
     capacity = db.Column(db.Integer)
     bed_quantity = db.Column(db.Integer)
